requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
redis==5.0.4
rq==1.16.1
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==6.1.0
python-dotenv==1.0.1
pydantic==2.7.4
pytest==8.2.0
httpx==0.27.0
